Move back to a single Scaffold per screen
Instead of two separate Scaffolds move back to a single Scaffold per
screen as having multiple Scaffold messed things up. E.g. the FAB was
behind the NavBar and some paddings weren't correct.
Still having a Scaffold per screen so actions, bottomBar, FAB and others
 can be configured per screen.
